pyproject.toml
[tool.poetry]
name = "test-case-junior-python-developer-abz-agency"
version = "0.1.0"
description = "Test case for the position of Junior Python Developer at abz.agency."
authors = ["sib_dip <tskdip@gmail.com>"]
license = "MIT"
readme = "README.md"
#packages = [{include = "test_case_junior_python_developer_abz_agency"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2 = "^2.9.5"
django-seed = "^0.3.1"
names-generator = "^0.1.0"
django-treebeard = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
